requirements.txt
psycopg2
reduct-py
